Migrate `deploy-tests` command
